Add File ID (UniqueID)
Hi there! I had used your blog post to set an image column but I noticed my images weren't showing up in PowerApps (desktop browser and mobile app) or Microsoft Lists (in iOS app). You can quickly recreate this issue by using your original script and loading the list in one of those apps (Lists iOS app is easiest so you don't have to create a PowerApp tied to your list). I dug in via REST APIs to look at the value of a column that was set via the UI and noticed an 'id' value in the Image JSON, and verified that it was in fact the UniqueId of the related file. This is why I've made the small modification above to get that ID and place it in the calls. Once you do that, your images will appear correctly in PowerApps and Microsoft Lists. If you don't do that, everything will look fine in SharePoint Online, but if you want to show the column in a PowerApp (or view in Lists iOS), your end users will report an issue that they're unable to see the images