Added Script to change retention labels on SP
This script reads SharePoint or OneDrive urls from a csv file, then goes through every item looking for specific retention labels. When they find matches, they change the labels to the new ones. The reason for this script is because you can't just change actions on existing retention labels beyond time period. Any major changes require the creation of a new label, and then setting it on every item where you need to replace the old one. This automates that.